\name{cat}
\alias{cat}
\title{JPEG picture of a cat}
\description{
JPEG file of a cat.
}
\usage{cat}
\format{JPEG}
\source{Original photograph taken by a member of rimage developer.}
\examples{
  data(cat)
  # Same as: x <- read.jpeg(system.file("data", "cat.jpg", package="rimage"))
  plot(cat)
}
\seealso{\code{\link{read.jpeg}}}
\keyword{datasets}


